Turn on more linters
